requirements.txt
# Flask Web Application Dependencies
Flask>=2.0.0
Flask-CORS>=3.0.0

# AI and Machine Learning
openai>=1.0.0
sentence-transformers>=2.0.0
faiss-cpu>=1.7.0
numpy>=1.21.0

# Data Processing
pandas>=1.3.0

# Additional utilities
requests>=2.25.0

# Optional: For production deployment
gunicorn>=20.0.0

# Install the dependencies when you deploy the code to any cloud environment
# Run: pip install -r requirements.txt
